chore: profile definition sources are read from configurable folder; * unit test skip loading of profile definition due to memory restrictions * performe some refactoring
