+++ .gitignore
@@ -36,3 +36,4 @@ out/
 ### VS Code ###
 .vscode/
 /node_modules/
+/deploy/.env
